+++ datasource/src/test/scala/astraea/spark/rasterframes/datasource/geotrellis/GeoTrellisDataSourceSpec.scala
@@ -22,6 +22,7 @@ import java.io.File
 import java.time.ZonedDateTime
 
 import astraea.spark.rasterframes._
+import astraea.spark.rasterframes.datasource.geotrellis.DefaultSource._
 import astraea.spark.rasterframes.util._
 import geotrellis.proj4.LatLng
 import geotrellis.raster._
@@ -41,7 +42,7 @@ import org.scalatest.{BeforeAndAfter, Inspectors}
 import org.apache.avro.generic._
 import org.apache.spark.storage.StorageLevel
 
-import scala.math.{min, max}
+import scala.math.{max, min}
 
 class GeoTrellisDataSourceSpec
     extends TestEnvironment with TestData with BeforeAndAfter with Inspectors
@@ -153,20 +154,23 @@ class GeoTrellisDataSourceSpec
   describe("DataSource options") {
     it("should respect partitions 2") {
       val expected = 2
-      val df = spark.read.option("numPartitions", expected)
-          .geotrellis.loadRF(layer)
+      val df = spark.read.geotrellis
+        .withNumPartitions(expected)
+        .loadRF(layer)
       assert(df.rdd.partitions.length === expected)
     }
     it("should respect partitions 20") {
       val expected = 20
-      val df = spark.read.option("numPartitions", expected)
-        .geotrellis.loadRF(layer)
+      val df = spark.read.geotrellis
+        .withNumPartitions(expected)
+        .loadRF(layer)
       assert(df.rdd.partitions.length === expected)
     }
     it("should respect split 2") {
       val param = 2
-      val df: RasterFrame = spark.read.option("tileSplits", param)
-        .geotrellis.loadRF(layer)
+      val df: RasterFrame = spark.read.geotrellis
+        .withTileSubdivisions(param)
+        .loadRF(layer)
 
       val dims = df.select(tileDimensions(df.tileColumns.head)("cols"), tileDimensions(df.tileColumns.head)("rows")).first()
       assert(dims.getAs[Int](0) === tileSize / param)
@@ -177,23 +181,25 @@ class GeoTrellisDataSourceSpec
     }
     it("should respect subdivide with TileFeature"){
       val param = 2
-      val rf: RasterFrame = spark.read.option("subdivideTile", param)
-        .geotrellis.loadRF(tfLayer)
-
-      assert(rf.count() === testRdd.count() * param * param)
+      val rf: RasterFrame = spark.read.geotrellis
+        .withTileSubdivisions(param)
+        .loadRF(tfLayer)
 
       val dims = rf.select(tileDimensions(rf.tileColumns.head)("cols"), tileDimensions(rf.tileColumns.head)("rows"))
         .first()
       assert(dims.getAs[Int](0) === tileSize / param)
       assert(dims.getAs[Int](1) === tileSize / param)
+
+      assert(rf.count() === testRdd.count() * param * param)
     }
     it("should respect both subdivideTile and numPartitions"){
       val subParam = 3
 
       val rf = spark.read
-        .option("subdivideTile", subParam)
-        .option("numPartitions", 7)
-        .geotrellis.loadRF(layer)
+        .geotrellis
+        .withNumPartitions(7)
+        .withTileSubdivisions(subParam)
+        .loadRF(layer)
 
       // is it partitioned correctly?
       assert(rf.rdd.partitions.length === 7)
@@ -209,13 +215,13 @@ class GeoTrellisDataSourceSpec
 
     it("should throw on subdivide 5") {
       // only throws when an action is taken...
-      assertThrows[IllegalArgumentException](spark.read.option("subdivideTile", 5).geotrellis.loadRF(layer).cache)
+      assertThrows[IllegalArgumentException](spark.read.geotrellis.withTileSubdivisions(5).loadRF(layer).cache)
     }
     it("should throw on subdivide 13") {
-      assertThrows[IllegalArgumentException](spark.read.option("subdivideTile", 13).geotrellis.loadRF(layer).cache)
+      assertThrows[IllegalArgumentException](spark.read.geotrellis.withTileSubdivisions(13).loadRF(layer).cache)
     }
     it("should throw on subdivide -3") {
-      assertThrows[IllegalArgumentException](spark.read.option("subdivideTile", -3).geotrellis.loadRF(layer).count)
+      assertThrows[IllegalArgumentException](spark.read.geotrellis.withTileSubdivisions(-3).loadRF(layer).count)
     }
   }
 
@@ -384,7 +390,7 @@ class GeoTrellisDataSourceSpec
     }
     it("should respect subdivideTile option on TileFeature RasterFrame") {
       val subParam = 4
-      val rf = spark.read.option("subdivideTile", subParam).geotrellis.loadRF(tfLayer)
+      val rf = spark.read.option(TILE_SUBDIVISIONS_PARAM, subParam).geotrellis.loadRF(tfLayer)
 
       assert(rf.count === testRdd.count * subParam * subParam)
 
@@ -397,8 +403,8 @@ class GeoTrellisDataSourceSpec
       val subParam = 2
 
       val rf = spark.read
-        .option("subdivideTile", subParam)
-        .option("numPartitions", 10)
+        .option(TILE_SUBDIVISIONS_PARAM, subParam)
+        .option(NUM_PARTITIONS_PARAM, 10)
         .geotrellis.loadRF(tfLayer)
 
       // is it subdivided?